Update logo image and reformat README
The logo image in the navigation bar has been replaced and resized for a better visual appearance. A major rewrite of the README.md file has been done to better explain the functionality of SymFlowBuilder, including its main features, its target audience, and its benefits in creating and managing Symfony workflows. This makes the README a more valuable resource for both new and existing users of SymFlowBuilder.