Add reactflow package and its dependencies to package-lock.json
The reactflow package along with several related dependencies have been added to the package-lock.json file. This addition is part of the ongoing work to modernize the codebase and provide better, smoother, and more efficient rendering of complex node and link diagrams in the application. The reactflow package specifically will provide new functionalities for diagram interactivity and design.