Refactor authentication options and refresh token path
This update simplifies the setup of the CredentialsProvider in NextAuth by removing unused credential options and streamlining the authorize callback to focus on email and password. It also incorporates a MongoDB connection within the authorize callback to find user records. Lastly, a dedicated /api/refresh-token endpoint is introduced to ensure efficient token refreshing.